requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
motor==3.3.2
httpx==0.27.0
anthropic==0.25.0
python-dotenv==1.0.1
